Mv zoom effect func to interactions
